+++ src/main/scala/com/serenity/session/SessionState.scala
@@ -663,6 +663,7 @@ given Encoder[AppConfig] = Encoder.instance { config =>
     "commandRunnerVisibleRows"          -> config.commandRunnerVisibleRows.asJson,
     "renderFpsTarget"                   -> config.renderFpsTarget.asJson,
     "editorInsertionTransitionKind"     -> config.editorInsertionTransitionKind.asJson,
+    "commandRunnerTransitionKind"       -> config.commandRunnerTransitionKind.asJson,
     "panelOpenTransitionKind"           -> config.panelOpenTransitionKind.asJson,
     "panelCloseTransitionKind"          -> config.panelCloseTransitionKind.asJson,
     "cursorMode"                        -> config.cursorMode.asJson,
@@ -730,11 +731,12 @@ given Decoder[AppConfig] = Decoder.instance { cursor =>
     editorInsertionTransitionKind <- cursor.getOrElse[TransitionKind]("editorInsertionTransitionKind")(
       TransitionKind.Fade
     )
-    panelOpenTransitionKind  <- cursor.getOrElse[Option[TransitionKind]]("panelOpenTransitionKind")(None)
-    panelCloseTransitionKind <- cursor.getOrElse[Option[TransitionKind]]("panelCloseTransitionKind")(None)
-    cursorMode               <- cursor.getOrElse[CursorMode]("cursorMode")(CursorMode.Blink)
-    cursorColors             <- cursor.getOrElse[CursorColorConfig]("cursorColors")(CursorColorConfig())
-    cursorInfoBarMode        <- cursor.getOrElse[CursorInfoBarMode]("cursorInfoBarMode")(CursorInfoBarMode.Off)
+    commandRunnerTransitionKind <- cursor.getOrElse[Option[TransitionKind]]("commandRunnerTransitionKind")(None)
+    panelOpenTransitionKind     <- cursor.getOrElse[Option[TransitionKind]]("panelOpenTransitionKind")(None)
+    panelCloseTransitionKind    <- cursor.getOrElse[Option[TransitionKind]]("panelCloseTransitionKind")(None)
+    cursorMode                  <- cursor.getOrElse[CursorMode]("cursorMode")(CursorMode.Blink)
+    cursorColors                <- cursor.getOrElse[CursorColorConfig]("cursorColors")(CursorColorConfig())
+    cursorInfoBarMode           <- cursor.getOrElse[CursorInfoBarMode]("cursorInfoBarMode")(CursorInfoBarMode.Off)
     cursorInfoBarPlacement <- cursor.getOrElse[CursorInfoBarPlacement]("cursorInfoBarPlacement")(
       CursorInfoBarPlacement.Floating
     )
@@ -773,6 +775,7 @@ given Decoder[AppConfig] = Decoder.instance { cursor =>
     commandRunnerVisibleRows = commandRunnerVisibleRows,
     renderFpsTarget = renderFpsTarget,
     editorInsertionTransitionKind = editorInsertionTransitionKind,
+    commandRunnerTransitionKind = commandRunnerTransitionKind,
     panelOpenTransitionKind = panelOpenTransitionKind,
     panelCloseTransitionKind = panelCloseTransitionKind,
     cursorMode = cursorMode,