+++ src/test/scala/com/serenity/ConfigManagerSpec.scala
@@ -574,6 +574,7 @@ class ConfigManagerSpec extends AnyFlatSpec with Matchers with OptionValues:
         |ui.motion.editor_text = typed
         |ui.motion.panel_open = directional
         |ui.motion.panel_close = off
+        |ui.motion.command_runner_reveal = outline
         |ui.motion.command_runner = subtle
         |ui.motion.ui = smooth
         |""".stripMargin
@@ -593,6 +594,7 @@ class ConfigManagerSpec extends AnyFlatSpec with Matchers with OptionValues:
     config.editorInsertionTransitionKind shouldBe TransitionKind.TypedText
     config.panelOpenTransitionKind shouldBe Some(TransitionKind.DirectionalSweep)
     config.panelCloseTransitionKind shouldBe Some(TransitionKind.Disabled)
+    config.commandRunnerTransitionKind shouldBe Some(TransitionKind.OutlineThenContent)
     config.commandRunnerAnimation shouldBe com.serenity.animation.AnimationConfig.subtle
     config.uiAnimation shouldBe com.serenity.animation.AnimationConfig.smooth
     config.characterAnimation shouldBe None
@@ -606,6 +608,7 @@ class ConfigManagerSpec extends AnyFlatSpec with Matchers with OptionValues:
     ConfigManager.configToString(config) should include("ui.motion.editor_text = typed")
     ConfigManager.configToString(config) should include("ui.motion.panel_open = directional")
     ConfigManager.configToString(config) should include("ui.motion.panel_close = off")
+    ConfigManager.configToString(config) should include("ui.motion.command_runner_reveal = outline")
     ConfigManager.configToString(config) should include("ui.motion.command_runner = subtle")
     ConfigManager.configToString(config) should include("ui.motion.ui = smooth")
   }